% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshevDistanceW.details.R
\name{chebyshevDistanceW.details}
\alias{chebyshevDistanceW.details}
\title{To calculate the Chebyshev distance applying weights.}
\usage{
chebyshevDistanceW.details(cluster1, cluster2, weight)
}
\arguments{
\item{cluster1}{is a cluster.}

\item{cluster2}{is a cluster.}

\item{weight}{is a numeric vector.}
}
\value{
Chebyshev distance applying weights value. Explanation.
}
\description{
To explain how to calculate the Chebyshev distance between clusters applying weights given.
}
\details{
The function calculates the Chebyshev distance value from \code{cluster1} and \code{cluster2}, applying weights to the cluster's components.
}
\examples{

cluster1 <- matrix(c(1,2),ncol=2)
cluster2 <- matrix(c(1,3),ncol=2)

weight1 <- c(0.4,0.6)
weight2 <- c(2,12)

chebyshevDistanceW.details(cluster1,cluster2,weight1)

chebyshevDistanceW.details(cluster1,cluster2,weight2)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
