% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDistanceByApproach.R
\name{clusterDistanceByApproach}
\alias{clusterDistanceByApproach}
\title{To calculate the distance by approach option.}
\usage{
clusterDistanceByApproach(distances, approach)
}
\arguments{
\item{distances}{is a numeric vector.}

\item{approach}{is a string. Type of function to apply.}
}
\value{
max, min or average from a vector.
}
\description{
To calculate the distance depending on \code{option} given.
}
\details{
This function is part of the hierarchical clusterization method. The function calculates the
distance value from \code{distances}.

\code{approach} indicates the algorithm used to get the value. Possible values: {\code{'MAX'},
\code{'MIN'}, \code{'AVG'}}.
}
\examples{

distances1 <- c(4,14,24,34)

distances2 <- c(1:10)

clusterDistanceByApproach(distances1,'MAX')

clusterDistanceByApproach(distances2,'MIN')

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
