% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCluster.details.R
\name{getCluster.details}
\alias{getCluster.details}
\title{To explain how to get the clusters with minimal distance.}
\usage{
getCluster.details(distance, matrix)
}
\arguments{
\item{distance}{is a number. It should be in the matrix.}

\item{matrix}{is a numeric matrix.}
}
\value{
Numeric vector with two clusters indexs.
}
\description{
To explain how to get the clusters with the minimal distance value. By using the given \code{distance},
it gets the matrix index.
}
\details{
This function is part of the hierarchical clusterization method. The function uses the
\code{distance} value and gets the clustersId with the minimal distance.

For the divisive algorithm, it chooses the distances from a distances list.
}
\examples{

matrixExample <- matrix(c(1:10), ncol=2)

getCluster.details(2,matrixExample)


}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
