% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterDivisive.R
\name{getClusterDivisive}
\alias{getClusterDivisive}
\title{To get the clusters with maximal distance.}
\usage{
getClusterDivisive(distance, vector)
}
\arguments{
\item{distance}{is a number. It should be in the matrix.}

\item{vector}{is a numeric vector}
}
\value{
A cluster.
}
\description{
To get the clusters with the maximal distance value. By using the given \code{distance},
it gets the matrix index.
}
\details{
This function is part of the hierarchical clusterization method. The function uses the
\code{distance} value and gets the clustersId with the minimal distance.

For the divisive algorithm, it chooses the distances from a distances list.
}
\examples{

getClusterDivisive(2,c(1:10))

getClusterDivisive(6,c(2,4,6,8,10,12))

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
