% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initData.R
\name{initData}
\alias{initData}
\title{To initialize data, hierarchical correlation algorithm.}
\usage{
initData(data)
}
\arguments{
\item{data}{is a data frame with the main data.}
}
\value{
A cluster list. Initializing data.
}
\description{
To initialize data, hierarchical correlation algorithm.
}
\details{
This function is part of the hierarchical correlation method. The function initializes \code{data} transforming each row from the data frame into
a matrix with every row elements.
}
\examples{

data <- matrix(c(1,2,1,4,5,1,8,2,9,6,3,5,8,5,4),ncol= 3)

dataFrame <- data.frame(data)

initData(dataFrame)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
