% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initImages.R
\name{initImages}
\alias{initImages}
\title{To display an image.}
\usage{
initImages(path)
}
\arguments{
\item{path}{is a file path.}
}
\description{
An auxiliar function to display a picture.
}
\details{
This function is part of the hierarchical clusterization method. The function calculates the
Euclidean distance value from \code{x} and \code{y}.
}
\examples{

path <- "../man/images/euclideanDistance.PNG"

\donttest{initImages(path)}


}
