% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDistance.R
\name{matrixDistance}
\alias{matrixDistance}
\title{Matrix distance by distance type}
\usage{
matrixDistance(list, distance)
}
\arguments{
\item{list}{is a clusters list.}

\item{distance}{is a literal.}
}
\description{
To calculate the matrix distance by using \code{distance} type.
}
\details{
This function is part of the hierarchical clusterization method. The function calculates the
matrix distance by using the distance type given.

The \code{list} parameter will be a list with the clusters as rows and columns.

The function avoids distances equal 0 and undefined clusters.
}
\examples{

data <- c(1:10)

clusters <- toList(data)

matrixDistance(clusters, 'EUC')
}
