% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDistance.R
\name{maxDistance}
\alias{maxDistance}
\title{Maximal distance}
\usage{
maxDistance(matrix)
}
\arguments{
\item{matrix}{is a numeric matrix. It could be a numeric vector.}
}
\value{
numeric value. Max value from a matrix
}
\description{
Get the matrix maximal value.
}
\details{
This function is part of the hierarchical clusterization method. The function uses the numeric
vector or matrix \code{matrix} given and return the maximal value.
The function avoids distances equal 0, and initialize maximal value with an auxiliar function \code{initMax},
which gets the first matrix element with a valid distance.
}
\examples{

matrixExample <- matrix(c(1:10), nrow=2)

maxDistance(1:10)

maxDistance(matrixExample)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
