% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minDistance.details.R
\name{minDistance.details}
\alias{minDistance.details}
\title{Minimal distance}
\usage{
minDistance.details(matrix)
}
\arguments{
\item{matrix}{is a numeric matrix. It could be a numeric vector.}
}
\value{
Numeric value. Min value from a matrix. Explanation.
}
\description{
To explain how to get the matrix minimal value.
}
\details{
This function is part of the hierarchical clusterization method. The function uses the numeric
vector or matrix \code{matrix} given and return the minimal value.
The function avoids distances equal 0, and initialize minimum value with an auxiliar function \code{initMin},
which gets the first matrix element with a valid distance.
}
\examples{

matrixExample <- matrix(c(1:10), nrow=2)

minDistance(1:10)

minDistance.details(matrixExample)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
