% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newCluster.R
\name{newCluster}
\alias{newCluster}
\title{To create a new cluster.}
\usage{
newCluster(list, clusters)
}
\arguments{
\item{list}{is the generic cluster list.}

\item{clusters}{is a vector with the matrix index clusters.}
}
\value{
A list with clusters.
}
\description{
To create the cluster formed by the two clusters given. Add the new cluster to \code{list}.
}
\details{
This function is part of the hierarchical clusterization method.

1 - The function maps \code{clusters} in \code{list}.

2 - It creates a new cluster from them.

3 - It adds the new cluster to \code{list}.

4 - It disables the clusters used in the second step.
}
\examples{

data <- c(1:10)

list <- toList(data)

clusters <- c(1,2)

newCluster(list,clusters)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
