% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octileDistance.R
\name{octileDistance}
\alias{octileDistance}
\title{To calculate the Octile distance.}
\usage{
octileDistance(x, y)
}
\arguments{
\item{x}{is a numeric vector or a matrix. It represents the values of a cluster.}

\item{y}{is a numeric vector or a matrix. It represents the values of a cluster.}
}
\value{
Octile distance value.
}
\description{
To calculate the octile distance of two clusters.
}
\details{
This function is part of the hierarchical clusterization method. The function calculates the
octile distance value from \code{x} and \code{y}.
}
\examples{

x <- c(1,2)
y <- c(1,3)

cluster1 <- matrix(x,ncol=2)
cluster2 <- matrix(y,ncol=2)

octileDistance(x,y)

octileDistance(cluster1,cluster2)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
