% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toList.R
\name{toList}
\alias{toList}
\title{To transform data into list}
\usage{
toList(data)
}
\arguments{
\item{data}{could be a numeric vector, a matrix or a numeric data frame.}
}
\value{
A list with clusters.
}
\description{
To transform \code{data} into list.
}
\details{
This function is part of the agglomerative hierarchical clusterization method. The function initializes
\code{data} content as a list.

In agglomerative algorithm, it adds a \code{TRUE} flag to each element, which indicates that the cluster is not grouped.
}
\examples{

data <- c(1:10)

matrix <- matrix(data,ncol=2)

dataFrame <- data.frame(matrix)

toList(data)

toList(matrix)

toList(dataFrame)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
