% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toListDivisive.details.R
\name{toListDivisive.details}
\alias{toListDivisive.details}
\title{To explain how to transform data into list}
\usage{
toListDivisive.details(data)
}
\arguments{
\item{data}{could be a numeric vector, a matrix or a numeric data frame.}
}
\value{
A list with clusters. Explanation.
}
\description{
To explain how to transform \code{data} into list.
}
\details{
This function is part of the divisive hierarchical clusterization method. The function initializes
\code{data} content as a list.
}
\examples{

data <- c(1:10)

matrix <- matrix(data,ncol=2)

dataFrame <- data.frame(matrix)

toListDivisive.details(data)

toListDivisive.details(matrix)

toListDivisive.details(dataFrame)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
