% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toListDivisive.R
\name{toListDivisive}
\alias{toListDivisive}
\title{To transform data into list}
\usage{
toListDivisive(data)
}
\arguments{
\item{data}{could be a numeric vector, a matrix or a numeric data frame.}
}
\value{
a list with clusters.
}
\description{
To transform \code{data} into list.
}
\details{
This function is part of the divisive hierarchical clusterization method. The function initializes
\code{data} content as a list.
}
\examples{

data <- c(1:10)

matrix <- matrix(data,ncol=2)

dataFrame <- data.frame(matrix)

toListDivisive(data)

toListDivisive(matrix)

toListDivisive(dataFrame)

}
\author{
Roberto Alcántara \email{roberto.alcantara@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
